+++ src/main/java/com/wedevol/springdatacouchbase/core/dao/PlaceRepository.java
@@ -1,0 +1,14 @@
+package com.wedevol.springdatacouchbase.core.dao;
+
+import org.springframework.data.repository.CrudRepository;
+import com.wedevol.springdatacouchbase.core.dao.doc.PlaceDoc;
+
+/**
+ * Standard CRUD repository for Place doc + query methods
+ * 
+ * @author Charz++
+ */
+
+public interface PlaceRepository extends CrudRepository<PlaceDoc, String> {
+
+}